Adds doctor variables to DLHS4 build #8
This adds two new variables to the DLHS build: one for contractual
general duty doctors and a second for regular general duty doctors. It
also includes a combined total doctors variable. These new variables
are more consistent with the doctor defintion in the PC build than the
original "total staff" variable in the DLHS build.