Migrate to new covid case data
We're using `covindia` case data now, not the data from
`covid19india.org`. Sunset the old components of the case data build
to reflect this change (and implement associated changes to the data
folder). Note that the `covindia` build directly matches to LGD
districts, rather than matching through PC11 and aggregating to LGD.